+++ src/main/java/com/example/bobfriend/controller/AuthenticationController.java
@@ -2,7 +2,7 @@ package com.example.bobfriend.controller;
 
 import com.example.bobfriend.jwt.JwtTokenProvider;
 import com.example.bobfriend.model.dto.MemberDto;
-import com.example.bobfriend.model.dto.TokenDto;
+import com.example.bobfriend.model.dto.token.*;
 import com.example.bobfriend.model.exception.MemberDuplicatedException;
 import com.example.bobfriend.service.AuthService;
 import lombok.RequiredArgsConstructor;
@@ -30,13 +30,13 @@ public class AuthenticationController {
     @PostMapping("/signin")
     public ResponseEntity signin(@Valid @RequestBody MemberDto.Login loginDto)
             throws AuthenticationException {
-        TokenDto.Token tokenDto = authService.signin(loginDto);
+        Token tokenDto = authService.signin(loginDto);
         return new ResponseEntity(tokenDto, HttpStatus.OK);
     }
 
 
     @PostMapping("/issue")
-    public ResponseEntity issueToken(@RequestBody TokenDto.Token token) {
+    public ResponseEntity issueToken(@RequestBody Token token) {
         return ResponseEntity.ok(authService.issueToken(token));
     }
 
@@ -44,7 +44,7 @@ public class AuthenticationController {
     @GetMapping("/validate")
     public ResponseEntity validateToken(HttpServletRequest request) {
         String token = tokenProvider.resolveToken(request);
-        return ResponseEntity.ok(new TokenDto.Validation(tokenProvider.validateAccessToken(token)));
+        return ResponseEntity.ok(new Validation(tokenProvider.validateAccessToken(token)));
     }
