+++ src/test/java/com/example/bobfriend/controller/AuthenticationControllerTest.java
@@ -2,7 +2,8 @@ package com.example.bobfriend.controller;
 
 import com.example.bobfriend.jwt.JwtTokenProvider;
 import com.example.bobfriend.model.dto.MemberDto;
-import com.example.bobfriend.model.dto.TokenDto;
+import com.example.bobfriend.model.dto.token.Token;
+import com.example.bobfriend.model.dto.token.Validation;
 import com.example.bobfriend.model.entity.Member;
 import com.example.bobfriend.model.entity.Sex;
 import com.example.bobfriend.service.AuthService;
@@ -75,7 +76,7 @@ public class AuthenticationControllerTest {
 
     @Test
     void signin() throws Exception {
-        TokenDto.Token tokenDto = new TokenDto.Token("jwt-access-token-example", "jwt-refresh-token-example");
+        Token tokenDto = new Token("jwt-access-token-example", "jwt-refresh-token-example");
         when(authService.signin(any())).thenReturn(tokenDto);
 
         MemberDto.Login login = MemberDto.Login.builder()
@@ -101,7 +102,7 @@ public class AuthenticationControllerTest {
 
     @Test
     void signup() throws Exception {
-        TokenDto.Token tokenDto = new TokenDto.Token("jwt-access-token-example", "jwt-refresh-token-example");
+        Token tokenDto = new Token("jwt-access-token-example", "jwt-refresh-token-example");
         MemberDto.Signup signup = MemberDto.Signup.builder()
                 .email(testMember.getEmail())
                 .nickname(testMember.getNickname())
@@ -145,7 +146,7 @@ public class AuthenticationControllerTest {
 
     @Test
     void reissueTest() throws Exception {
-        TokenDto.Token tokenDto = TokenDto.Token.builder()
+        Token tokenDto = Token.builder()
                 .accessToken("new-access-token-example")
                 .refreshToken("new-refresh-token-example")
                 .build();
@@ -153,7 +154,7 @@ public class AuthenticationControllerTest {
                 .thenReturn(tokenDto);
         mvc.perform(post("/api/issue")
                         .content(objectMapper.writeValueAsString(
-                                TokenDto.Token.builder()
+                                Token.builder()
                                         .accessToken("old-access-token")
                                         .refreshToken("old-refresh-token")
                                         .build()
@@ -178,7 +179,7 @@ public class AuthenticationControllerTest {
                 .andExpect(status().isOk())
                 .andExpect(content().json(
                         objectMapper.writeValueAsString(
-                                new TokenDto.Validation(tokenProvider.validateAccessToken(any()))
+                                new Validation(tokenProvider.validateAccessToken(any()))
                         )
                 ))
                 .andDo(document("auth/validate",
